requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain_community>=0.0.13
langchain-google-genai>=0.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10
pymupdf>=1.23.7
docx2txt>=0.8
pandas>=2.0.3
openpyxl>=3.1.2
python-magic>=0.4.27
tiktoken>=0.5.1
protobuf>=4.24.4
google-generativeai>=0.3.2
matplotlib>=3.7.0
plotly>=5.14.0
wordcloud>=1.8.2
nltk>=3.8.1

# Requirements for local models
torch>=2.0.0
transformers>=4.35.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
einops>=0.7.0
xformers>=0.0.22
